--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -2,19 +2,17 @@ using Stencils
 using Aqua
 using Test
 
-if VERSION >= v"1.9.0"
-    @testset "Aqua.jl" begin
-        Aqua.test_all(
-            Stencils;
-            ambiguities=(exclude=[Base.copy!, Base.copyto!],),
-            unbound_args=true,
-            undefined_exports=true,
-            project_extras=true,
-            stale_deps=true,
-            deps_compat=true,
-            piracies=true,
-        )
-    end
+@testset "Aqua.jl" begin
+    Aqua.test_all(
+        Stencils;
+        ambiguities=(exclude=[Base.copy!, Base.copyto!],),
+        unbound_args=true,
+        undefined_exports=true,
+        project_extras=true,
+        stale_deps=true,
+        deps_compat=true,
+        piracies=true,
+    )
 end
 
 include("stencils.jl")