@evetion This should be a start at least, see what the tests do...

Closes #56 and closes #62