@rafaqz 2025-03-29
@evetion this should be working, could you test behaviour and performance in Geomorphometry.jl?