@evetion 2025-03-29
Hmm, not sure where my comment went. Here goes again:

This branch gives me speedups when testing TPI with a random matrix:

---

## new
10x10 matrix
```
julia> @benchmark Geomorphometry.TPI($A)
BenchmarkTools.Trial: 10000 samples with 8 evaluations per sample.
 Range (min … max):  3.042 μs …   9.245 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.182 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.277 μs ± 286.733 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▃▇██▇▆▆▆▆▆▅▅▄▄▄▃▃▃▂▃▂▂▁▁▁▁▁▁ ▁                              ▂
  ████████████████████████████████▇███▇██▇▇█▆█▇▅▆▇▆█▆▇▆▆▄▇▆▆▇ █
  3.04 μs      Histogram: log(frequency) by time      4.43 μs <

 Memory estimate: 2.28 KiB, allocs estimate: 29.
```

100x100 matrix
```
julia> @benchmark Geomorphometry.TPI($A)
BenchmarkTools.Trial: 10000 samples with 1 evaluation per sample.
 Range (min … max):  32.541 μs …  12.370 ms  ┊ GC (min … max): 0.00% … 99.44%
 Time  (median):     37.583 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   41.838 μs ± 170.400 μs  ┊ GC (mean ± σ):  5.74% ±  1.41%

        ▄█▆▆▃
  ▃▂▂▂▂███████▆▅▅▄▄▄▄▃▃▃▃▃▃▃▃▃▃▃▃▂▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂ ▃
  32.5 μs         Histogram: frequency by time         63.9 μs <

 Memory estimate: 79.56 KiB, allocs estimate: 30.
```


## old
10x10 matrix
```
julia> @benchmark TPI($A)
BenchmarkTools.Trial: 10000 samples with 5 evaluations per sample.
 Range (min … max):  6.750 μs …  17.717 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     7.008 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   7.186 μs ± 593.816 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▂▆██▇▇▇▆▆▅▅▄▄▃▃▃▂▂▁▁▁▁   ▁                                  ▂
  ███████████████████████████▇██▇███▇▇▇▇▇█▇▇▇▇▇█▆▆▅▆▅▆▆▆▄▅▂▃▄ █
  6.75 μs      Histogram: log(frequency) by time      9.59 μs <

 Memory estimate: 4.17 KiB, allocs estimate: 71.
```
100x100 matrix
```
julia> @benchmark TPI($A)
BenchmarkTools.Trial: 10000 samples with 1 evaluation per sample.
 Range (min … max):  40.583 μs …  12.778 ms  ┊ GC (min … max): 0.00% … 99.38%
 Time  (median):     46.000 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   53.230 μs ± 229.611 μs  ┊ GC (mean ± σ):  9.73% ±  2.42%

       ▃█▅▁
  ▃▂▂▄█████▆▅▅▄▄▄▃▃▃▄▄▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▂▂▂▂▁▂▂▂▁▁▂▂▁▂▂ ▃
  40.6 μs         Histogram: frequency by time         83.6 μs <

 Memory estimate: 81.45 KiB, allocs estimate: 72.
```

Overall still a bit slower than LocalFilters on small matrices, but the break-even point comes sooner, around 25x25.

---

That said, I still see repeated compilation:

new
```
julia> @time Geomorphometry.TPI(A)
  0.000067 seconds (29 allocations: 2.281 KiB, 0.81% compilation time)

julia> @time Geomorphometry.TPI(A)
  0.000051 seconds (29 allocations: 2.281 KiB, 1.14% compilation time)
```


old
```julia
julia> @time TPI(A);
  0.000086 seconds (71 allocations: 4.172 KiB, 1.07% compilation time)

julia> @time TPI(A);
  0.000092 seconds (71 allocations: 4.172 KiB, 1.36% compilation time)
```


Definition of TPI so you can reproduce
```julia
_tpi(x) = center(x) - mean(x)
function TPI(dem::AbstractMatrix{<:Real}, window::Stencil = Moore(1))
    mapstencil(_tpi, window, dem)
end
```