@rafaqz 2025-03-29
Huh. I guess KernelAbstracts and threading adds some weight to the small windows?

But it shouldnt be unstable anywhere now, so something could still be improved