@rafaqz 2025-03-29
Ok I've overhauled type stability, there were some problems that were not there in earlier Julia version. KA still has a tiny bit of type stability problems internally. But we should be good for small arrays now.

For 10x10 Ive got 311 ns vs your  6.75 μs  above:
```julia
BenchmarkTools.Trial: 10000 sampleas with 236 evaluations per sample.
 Range (min … max):  311.017 ns … 317.908 μs  ┊ GC (min … max):  0.00% … 99.75%
 Time  (median):     341.102 ns               ┊ GC (median):     0.00%
 Time  (mean ± σ):   529.877 ns ±   3.348 μs  ┊ GC (mean ± σ):  18.50% ±  7.13%

  ▆█▅▃▁   ▄▄▂▁  ▁▂▃▃▄▃▁▁▁  ▁▁                                   ▁
  ██████▇█████████████████████▇▇▇▇▆▆▇▇▆▆▆▆▆▅▆▆▆▄▅▅▄▄▅▆▄▄▄▄▃▅▂▄▅ █
  311 ns        Histogram: log(frequency) by time       1.25 μs <

 Memory estimate: 1.22 KiB, allocs estimate: 12.
```