@evetion 2025-03-29
That's fantastic, can confirm its even faster now (and beats LocalFilters). It also doesn't compile anymore. ❤️